Proyecto/requirements.txt
Django==5.1
django-rest-framework==0.1.0
drf-spectacular==0.27.2
PyJWT[crypto]==2.8.0
cryptography==41.0.3
djangorestframework-simplejwt==5.2.2
requests==2.31.0